Moved test data into test classes that use it.
